raspberrypizero/requirements.txt
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
colorzero==2.0
distro==1.8.0
getrpimodel==0.1.25
gpiozero==2.0
idna==3.3
lgpio==0.2.2.0
mh_z19==3.1.5
paho-mqtt==2.1.0
pigpio==1.78
pillow==10.3.0
pyserial==3.5
python-apt==2.6.0
requests==2.28.1
RPi.GPIO==0.7.1a4
six==1.16.0
smbus2==0.4.2
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
urllib3==1.26.12